requirements.txt
anthropic>=0.40.0
neo4j>=5.20.0
feedparser>=6.0.0
aiohttp>=3.9.0
pydantic>=2.5.0
python-dotenv>=1.0.0
lxml>=5.0.0
